[TNT-181] refactor: Chip 컴포넌트 이름 수정 및 ChipStyle 추가
