--- core/designsystem/src/main/java/co/kr/tnt/designsystem/component/tag/model/TagType.kt
@@ -1,6 +1,0 @@
-package co.kr.tnt.designsystem.component.tag.model
-
-enum class TagType {
-    EXERCISE,
-    DIET,
-}
